Created SQLiteHand;er module
